requirements.txt
# Cash Flow Dashboard - Dependencies
# Core Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Data Validation
pandera>=0.18.0
pydantic>=2.0.0

# Date/Time Handling
python-dateutil>=2.8.0

# Dashboard & Visualization
streamlit>=1.30.0
plotly>=5.18.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0
